Move build-related JS to resources/js-build
